requirements.txt
torch>=2.6.0
transformers>=4.36.0
datasets>=2.12.0
Pillow>=10.3.0
tqdm>=4.65.0
